fix(mobile): move auth components out of app/ directory
Files inside app/ are registered as routes by Expo Router, which caused
a crash when navigating to the Measure tab. Moved email-login-step,
otp-verify-step, oauth-icons, and offline-banner from
app/(auth)/components/ to components/auth/.
