+++ apps/mobile/src/components/auth/offline-banner.tsx
